refactor: move server initialization to app.js for better modularity
